support both perf APIs
Most current distributions support a version of perf with an older,
and slightly different API for the event handling routines.  The older
API does not have a final "perf_sample_dict" parameter.

If `splat` is used with a version of perf with a different API,
then an error occurs:
  TypeError: sdt_libpthread__mutex_entry_1_new() takes exactly 11 arguments (10 given)

A new command line parameter, "--api=<n>" is added by this patch.

To use splat with the older perf API:

$ ./splat.py -- --api=1

To use splat with the newer perf API:

$ ./splat.py -- --api=2

The default behavior is to use the older perf API, so that
splat.py can be used easily on current Linux distributions.

Fixes #39

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>