track trylocks
Trylock functions are not instrumented in GLibC.

It is still possible to track them using perf function probes.

Including the `mutex` parameter in the recorded data requires
the glilbc-debuginfo packages to be installed.  (Or, possibly
and architecture-specific register, which is challenging.)

Tracking the success or failure of the trylock call requires
using a return probe and capturing the return value, but the IP
won't match the IP of the trylock entry.  So, save this information
within the task instance, and match it up later.

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>